provide custom parsing functions to options
